requirements.txt
# Core Computer Vision and Numerical Computing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0

# 3D Processing and Mesh Generation
open3d>=0.17.0
trimesh>=3.20.0
scikit-learn>=1.3.0

# Image Processing and Visualization
matplotlib>=3.7.0
Pillow>=9.5.0

# Configuration Management
PyYAML>=6.0

# Testing Framework
pytest>=7.4.0
hypothesis>=6.82.0
pytest-cov>=4.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Deep Learning (for future RAFT-Stereo integration)
# torch>=2.0.0
# torchvision>=0.15.0

# Optional: Jupyter for analysis and visualization
# jupyter>=1.0.0
# ipywidgets>=8.0.0